+++ src/WarmUp/Sources/Warmup/WarmUp.swift
@@ -1,0 +1,6 @@
+import Foundation
+import Numerics
+
+func sum(_ x: Int, _ y: Int) -> Int {
+  return x + y
+}